chore: moved Available enum to its own file
