+++ Tests/AdyenTests/Adyen Tests/Components/Apple Pay/ApplePayComponentTests.swift
@@ -282,21 +282,3 @@ class ApplePayComponentTest: XCTestCase {
         return networks
     }
 }
-
-enum Available {
-    static var iOS12: Bool {
-        if #available(iOS 12.0, *) {
-            return true
-        } else {
-            return false
-        }
-    }
-    
-    static var iOS13: Bool {
-        if #available(iOS 13.0, *) {
-            return true
-        } else {
-            return false
-        }
-    }
-}
